pyproject.toml
[project]
name = "video-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "moviepy>=2.1.2",
    "openai>=1.78.0",
    "python-dotenv>=1.1.0",
    "python-ffmpeg>=2.0.12",
]
